refactor: Changes to initially though structure, guided by internal meeting discussion and US changes
