+++ frontend/folder-structure.md
@@ -1,0 +1,41 @@
+- **Root Directory:**
+  - `dist/`: Output directory for compiled and optimized code. Generated when you run the build command, (in this case, `pnpm build` ) in a Vite project. The contents of this directory are what you would deploy to a web server or hosting service.
+  - `index.html`: Main HTML file. Here the "root" div is the target for replacing content during the build process. The main module is loaded into this HTML file. It's the entry point of your application.
+  - `node_modules/`: This directory contains all the project's dependencies. It's generated and managed by the package manager (in this case, pnpm) based on the information in package.json.
+  - `package.json`: The package manager configuration file. It contains metadata about the project and its dependencies, as well as scripts to run various tasks, like building the project or starting a development server.
+  - `pnpm-lock.yaml`: This is the configuration file for `pnpm`. It contains information about the exact versions of dependencies installed, similar to `package-lock.json` used by npm or `yarn.lock` used by Yarn.
+  - `src/`: The source code directory.
+  - `tsconfig.json`: The TypeScript configuration file. It contains settings and options for the TypeScript compiler.
+  - `tsconfig.node.json`: A TypeScript configuration file specifically for Node.js.
+  - `vite.config.ts`: The Vite configuration file. It allows you to customize the behavior of Vite during development and production. This file can include settings for plugins, server configurations, and more.
+  - `.gitignore`: Specifies intentionally untracked files that Git should ignore, typically including `node_modules/`, `dist/`, and other temporary files.
+  - `.eslintrc.cjs`: This file is the configuration file for ESLint, a linter tool for identifying and reporting on patterns found in ECMAScript/JavaScript code.
+
+- **src/:**
+  - `assets/`: Directory for storing static assets used in the project.
+    - `icons/`: Folder to store icons.
+  - `components/`: Directory containing modular, reusable building blocks that encapsulate specific UI elements or functionalities. 
+    - `ui/`: Subdirectory for components dedicated to user interface.
+      - `{Broader-Category|Component}(example: Buttons)/`: Subdirectory category x components.
+        - `{Component}(example: DefaultButton)/`: Subdirectory for the component.
+          - `{Component}.styles.ts`: Styles for the default component.
+          - `{Component}.tsx`: Implementation of the component.
+  - `features/`: Directory containing specialized components that encapsulate specific features or functionalities of the application.
+    - `{Broader-Category|Feature}(example: Counter)/`: Subdirectory for a counter feature.
+      - `{Feature}.styles.ts`: Styles for the feature.
+      - `{Feature}.tsx`: Implementation of the feature.
+  - `pages/`: Directory containing structures and layouts of a specific page in the application, associated with routes.
+    - `{Page}(example: Home)/`: Subdirectory for the page component.
+      - `{Page}.styles.ts`: Styles for the page.
+      - `{Page}.tsx`: Implementation of the page.
+  - `modules/`: Directory for business logic modules, separated by responsibility.
+    - `{Module}(example: User)/`: Subdirectory for the user module.
+      - `client.ts`: Client module handling calls related to the module.
+      - `service.ts`: Service module responsible for calling clients, passing params, and applying some logic.
+      - `mapper.ts`: Mapper module transforming service responses to types and vice versa.
+      - `types.ts`: Types module representing data in the way it's known by the app.
+  
+  - `App.tsx`: The main React component file. This is where the structure and behavior of the main application are defined.
+  - `main.tsx`: Main TypeScript file. It might include the initialization and configuration of the React application.
+  - `vite-env.d.ts`: TypeScript declaration file for Vite-specific environment variables.
+
