--- frontend/public/index.html
@@ -1,13 +1,0 @@
-<!-- public/index.html -->
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>Your React App</title>
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
-</html>
