+++ frontend/src/App.tsx
@@ -1,0 +1,9 @@
+import HomePage from './pages/Home/Home'
+
+function App() {
+  return (
+    <HomePage/>
+  )
+}
+
+export default App
