--- frontend/src/components/Buttons/DefaultButton/DefaultButton.styles.ts
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const Button = styled('button')(({ theme }) => ({
-  backgroundColor: 'blue',
-  margin: '16px'
-}));
