--- frontend/src/components/Buttons/DefaultButton/index.ts
@@ -1,1 +1,0 @@
-export { default } from './DefaultButton';
