+++ frontend/src/components/ui/Buttons/DefaultButton/DefaultButton.styles.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Button = styled.button({
+    backgroundColor: 'blue',
+    margin: '16px'
+  });
