--- frontend/src/features/Counter/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Counter';
