+++ frontend/src/pages/Home/Home.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import Counter from '../../features/Counter';
+import Counter from '../../features/Counter/Counter';
 
 const HomePage: React.FC = () => {
   return (
     <div>
-      <h1>Welcome to My App</h1>
+      <h1>Skill-Track front draft</h1>
       <Counter />
     </div>
   );
